[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.05 ms, Memory: 75.1 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,8 +1,8 @@
 class Solution {
-    public long[] solution(int x, int n) {
+    public long[] solution(long x, int n) {
         long[] arr = new long[n];
         for (int i = 0 ; i < n ; i++) {
-            arr[i] = (long) x * (i + 1);
+            arr[i] = x * (i + 1);
         }
     return arr;
     }
